chore: reorganize classes (#61)
* chore: reorganize classes

* Remove unused listener

* Undo minified enabled

* Linting
