--- unleashandroidsdk/src/main/java/io/getunleash/android/ReadyListener.kt
@@ -1,5 +1,0 @@
-package io.getunleash.android
-
-fun interface ReadyListener {
-    fun onReady(): Unit
-}
