--- unleashandroidsdk/src/main/java/io/getunleash/android/cache/ToggleStoreListener.kt
@@ -1,10 +1,0 @@
-package io.getunleash.android.cache
-
-import io.getunleash.android.data.Toggle
-
-fun interface ToggleStoreListener {
-    /**
-     * This method will be called after the feature toggle store (usually a cache) is updated.
-     */
-    fun onTogglesStored(flags: Map<String, Toggle>)
-}
