+++ unleashandroidsdk/src/main/java/io/getunleash/android/data/Toggle.kt
@@ -5,7 +5,7 @@ package io.getunleash.android.data
  * For creating toggles see docs - [Feature toggles](https://docs.getunleash.io/docs/user_guide/create_feature_toggle)
  * @property name Name of the toggle
  * @property enabled Did this toggle get evaluated to true
- * @property variant used by [io.getunleash.UnleashClientSpec.getVariant] to get the variant data
+ * @property variant used by [io.getunleash.android.Unleash.getVariant] to get the variant data
  */
 data class Toggle(
     val name: String,
