+++ unleashandroidsdk/src/main/java/io/getunleash/android/errors/NotAuthorizedException.kt
@@ -1,4 +1,3 @@
 package io.getunleash.android.errors
 
-class NotAuthorizedException : Exception("Not authorized") {
-}
+class NotAuthorizedException : Exception("Not authorized")
