+++ unleashandroidsdk/src/main/java/io/getunleash/android/errors/ServerException.kt
@@ -1,4 +1,3 @@
 package io.getunleash.android.errors
 
-class ServerException(statusCode: Int) : Exception("Unleash responded with $statusCode") {
-}
+class ServerException(statusCode: Int) : Exception("Unleash responded with $statusCode")
