+++ unleashandroidsdk/src/main/java/io/getunleash/android/events/HeartbeatEvent.kt
@@ -1,6 +1,6 @@
 package io.getunleash.android.events
 
-import io.getunleash.android.data.Status
+import io.getunleash.android.polling.Status
 
 
 /**
