--- unleashandroidsdk/src/main/java/io/getunleash/android/events/UnleashAllEventsListener.kt
@@ -1,22 +1,0 @@
-package io.getunleash.android.events
-
-import io.getunleash.android.data.ImpressionEvent
-
-interface UnleashListener
-interface UnleashReadyListener: UnleashListener {
-    fun onReady()
-}
-
-interface UnleashStateListener: UnleashListener {
-    fun onStateChanged()
-}
-
-interface UnleashImpressionEventListener: UnleashListener {
-    fun onImpression(event: ImpressionEvent)
-}
-
-interface UnleashFetcherHeartbeatListener: UnleashListener {
-    fun onError(event: HeartbeatEvent)
-    fun togglesChecked()
-    fun togglesUpdated()
-}
