+++ unleashandroidsdk/src/main/java/io/getunleash/android/events/UnleashFetcherHeartbeatListener.kt
@@ -1,0 +1,7 @@
+package io.getunleash.android.events
+
+interface UnleashFetcherHeartbeatListener: UnleashListener {
+    fun onError(event: HeartbeatEvent)
+    fun togglesChecked()
+    fun togglesUpdated()
+}
