+++ unleashandroidsdk/src/main/java/io/getunleash/android/events/UnleashImpressionEventListener.kt
@@ -1,0 +1,7 @@
+package io.getunleash.android.events
+
+import io.getunleash.android.data.ImpressionEvent
+
+interface UnleashImpressionEventListener: UnleashListener {
+    fun onImpression(event: ImpressionEvent)
+}
