+++ unleashandroidsdk/src/main/java/io/getunleash/android/events/UnleashListener.kt
@@ -1,0 +1,4 @@
+package io.getunleash.android.events
+
+interface UnleashListener
+
