+++ unleashandroidsdk/src/main/java/io/getunleash/android/events/UnleashReadyListener.kt
@@ -1,0 +1,5 @@
+package io.getunleash.android.events
+
+interface UnleashReadyListener: UnleashListener {
+    fun onReady()
+}
