+++ unleashandroidsdk/src/main/java/io/getunleash/android/events/UnleashStateListener.kt
@@ -1,0 +1,5 @@
+package io.getunleash.android.events
+
+interface UnleashStateListener: UnleashListener {
+    fun onStateChanged()
+}
