+++ unleashandroidsdk/src/main/java/io/getunleash/android/polling/Status.kt
@@ -1,0 +1,13 @@
+package io.getunleash.android.polling
+
+enum class Status {
+    SUCCESS,
+    NOT_MODIFIED,
+    FAILED,
+    THROTTLED;
+
+    fun isSuccess() = this == SUCCESS
+
+    fun isNotModified() = this == NOT_MODIFIED
+    fun isFailed() = this == FAILED || this == THROTTLED
+}
